Added Multi-Tenancy

+++ Servers/utils/autoDriver.utils.ts
@@ -2,9 +2,9 @@ import { Transaction } from "sequelize";
 import { sequelize } from "../database/db";
 import { VendorModel } from "../domain.layer/models/vendor/vendor.model";
 
-export async function getData(tableName: string, transaction: Transaction) {
+export async function getData(tableName: string, tenant: string, transaction: Transaction) {
   const result = await sequelize.query(
-    `SELECT * FROM ${tableName} WHERE is_demo;`,
+    `SELECT * FROM "${tenant}".${tableName} WHERE is_demo;`,
     { transaction }
   );
   return result[0];
@@ -18,9 +18,9 @@ export async function insertData(
   return result;
 }
 
-export async function deleteDemoVendorsData(transaction: Transaction) {
+export async function deleteDemoVendorsData(tenant: string, transaction: Transaction) {
   const result = await sequelize.query(
-    `SELECT id FROM vendors WHERE is_demo;`,
+    `SELECT id FROM "${tenant}".vendors WHERE is_demo;`,
     {
       mapToModel: true,
       model: VendorModel,
@@ -31,7 +31,7 @@ export async function deleteDemoVendorsData(transaction: Transaction) {
   await Promise.all(
     result.map(async (r) => {
       await sequelize.query(
-        `DELETE FROM vendors_projects WHERE vendor_id = :vendor_id`,
+        `DELETE FROM "${tenant}".vendors_projects WHERE vendor_id = :vendor_id`,
         {
           replacements: { vendor_id: r.id },
           transaction,
@@ -42,7 +42,7 @@ export async function deleteDemoVendorsData(transaction: Transaction) {
   await Promise.all(
     result.map(async (r) => {
       await sequelize.query(
-        `DELETE FROM vendorrisks WHERE vendor_id = :vendor_id`,
+        `DELETE FROM "${tenant}".vendorrisks WHERE vendor_id = :vendor_id`,
         {
           replacements: { vendor_id: r.id },
           transaction,
@@ -50,5 +50,5 @@ export async function deleteDemoVendorsData(transaction: Transaction) {
       );
     })
   );
-  await sequelize.query(`DELETE FROM vendors WHERE is_demo;`, { transaction });
+  await sequelize.query(`DELETE FROM "${tenant}".vendors WHERE is_demo;`, { transaction });
 }
