+++ Servers/utils/projectScope.utils.ts
@@ -3,9 +3,11 @@ import { sequelize } from "../database/db";
 import { QueryTypes, Transaction } from "sequelize";
 import { IProjectScope } from "../domain.layer/interfaces/i.projectScope";
 
-export const getAllProjectScopesQuery = async (): Promise<IProjectScope[]> => {
+export const getAllProjectScopesQuery = async (
+  tenant: string
+): Promise<IProjectScope[]> => {
   const projectScopes = await sequelize.query(
-    "SELECT * FROM projectscopes ORDER BY created_at DESC, id ASC",
+    `SELECT * FROM "${tenant}".projectscopes ORDER BY created_at DESC, id ASC`,
     {
       mapToModel: true,
       model: ProjectScopeModel,
@@ -15,10 +17,11 @@ export const getAllProjectScopesQuery = async (): Promise<IProjectScope[]> => {
 };
 
 export const getProjectScopeByIdQuery = async (
-  id: number
+  id: number,
+  tenant: string
 ): Promise<IProjectScope | null> => {
   const result = await sequelize.query(
-    "SELECT * FROM projectscopes WHERE id = :id",
+    `SELECT * FROM "${tenant}".projectscopes WHERE id = :id`,
     {
       replacements: { id },
       mapToModel: true,
@@ -30,10 +33,11 @@ export const getProjectScopeByIdQuery = async (
 
 export const createProjectScopeQuery = async (
   projectScope: Partial<ProjectScopeModel>,
+  tenant: string,
   transaction: Transaction
 ): Promise<ProjectScopeModel> => {
   const result = await sequelize.query(
-    `INSERT INTO projectscopes (
+    `INSERT INTO "${tenant}".projectscopes (
       assessment_id, describe_ai_environment, is_new_ai_technology,
       uses_personal_data, project_scope_documents, technology_type,
       has_ongoing_monitoring, unintended_outcomes, technology_documentation
@@ -66,6 +70,7 @@ export const createProjectScopeQuery = async (
 export const updateProjectScopeByIdQuery = async (
   id: number,
   projectScope: Partial<ProjectScopeModel>,
+  tenant: string,
   transaction: Transaction
 ): Promise<ProjectScopeModel | null> => {
   const updateProjectScope: Partial<Record<keyof ProjectScopeModel, any>> = {};
@@ -93,7 +98,7 @@ export const updateProjectScopeByIdQuery = async (
     .map((f) => `${f} = :${f}`)
     .join(", ");
 
-  const query = `UPDATE projectscopes SET ${setClause} WHERE id = :id RETURNING *;`;
+  const query = `UPDATE "${tenant}".projectscopes SET ${setClause} WHERE id = :id RETURNING *;`;
 
   updateProjectScope.id = id;
 
@@ -110,10 +115,11 @@ export const updateProjectScopeByIdQuery = async (
 
 export const deleteProjectScopeByIdQuery = async (
   id: number,
+  tenant: string,
   transaction: Transaction
 ): Promise<Boolean> => {
   const result = await sequelize.query(
-    "DELETE FROM projectscopes WHERE id = :id RETURNING *",
+    `DELETE FROM "${tenant}".projectscopes WHERE id = :id RETURNING *`,
     {
       replacements: { id },
       mapToModel: true,
